+++ CREATE_TABLE_DDL/010.household_demographics.sql
@@ -1,6 +1,6 @@
 CREATE TABLE household_demographics (
-    hd_demo_sk integer NOT NULL,
-    hd_income_band_sk integer,
+    hd_demo_sk bigint NOT NULL,
+    hd_income_band_sk bigint,
     hd_buy_potential character(15),
     hd_dep_count integer,
     hd_vehicle_count integer
