+++ CREATE_TABLE_DDL/011.income_band.sql
@@ -1,5 +1,5 @@
 CREATE TABLE income_band (
-    ib_income_band_sk integer NOT NULL,
+    ib_income_band_sk bigint NOT NULL,
     ib_lower_bound integer,
     ib_upper_bound integer
 )
