+++ CREATE_TABLE_DDL/012.inventory.sql
@@ -1,6 +1,6 @@
 CREATE TABLE inventory (
-    inv_date_sk integer NOT NULL,
-    inv_item_sk integer NOT NULL,
-    inv_warehouse_sk integer NOT NULL,
+    inv_date_sk bigint NOT NULL,
+    inv_item_sk bigint NOT NULL,
+    inv_warehouse_sk bigint NOT NULL,
     inv_quantity_on_hand integer
 )
